refactor: 내 정보를 수정하는 함수를 내 정보 페이지로 이동
  - 기존에는 비밀번호 확인 컴포넌트에 있었으나, 해당 컴포넌트의 역할과는 맞지 않다고 판단하여 이동
