+++ frontend/src/components/ConfirmPassword/index.tsx
@@ -1,75 +1,24 @@
 import React, { useRef } from 'react';
 
-import { useRecoilState, useRecoilValue } from 'recoil';
+import { useRecoilValue } from 'recoil';
 
-import {
-  getUserInfo,
-  requestChangeName,
-  requestChangePassword,
-  requestConfirmPassword,
-} from 'apis/request/user';
+import { requestConfirmPassword } from 'apis/request/user';
 import Modal from 'components/@shared/Modal';
-import { ERROR_MESSAGE, GUIDE_MESSAGE } from 'constants/message';
+import { ERROR_MESSAGE } from 'constants/message';
 import useModal from 'hooks/useModal';
-import useSnackbar from 'hooks/useSnackbar';
-import { loginState, modalState } from 'store/states';
+import { modalState } from 'store/states';
 
 import * as S from './index.styled';
 
 interface ConfirmPasswordProps {
-  type: string;
-  newValue: string;
-  setIsEditable: React.Dispatch<React.SetStateAction<boolean>>;
+  editValue: () => void;
 }
 
-function ConfirmPassword({
-  type,
-  newValue,
-  setIsEditable,
-}: ConfirmPasswordProps) {
-  const [loginInfo, setLoginInfo] = useRecoilState(loginState);
-
+function ConfirmPassword({ editValue }: ConfirmPasswordProps) {
   const modalFlag = useRecoilValue(modalState);
   const { setOffModal } = useModal();
 
-  const { setMessage } = useSnackbar();
-
   const confirmPasswordRef = useRef<HTMLInputElement>(null);
-
-  const editValue = () => {
-    switch (type) {
-      case 'name':
-        requestChangeName(newValue)
-          .then(() => {
-            setMessage(GUIDE_MESSAGE.MEMBER.SUCCESS_NAME_REQUEST);
-            setIsEditable(false);
-
-            getUserInfo().then(userInfo => {
-              setLoginInfo({ ...loginInfo, user: userInfo });
-            });
-          })
-          .catch(() => {
-            alert(ERROR_MESSAGE.MEMBER.FAILURE_NAME_REQUEST);
-          });
-
-        break;
-      case 'password':
-        requestChangePassword(newValue)
-          .then(() => {
-            setMessage(GUIDE_MESSAGE.MEMBER.SUCCESS_PASSWORD_REQUEST);
-            setIsEditable(false);
-
-            getUserInfo().then(userInfo => {
-              setLoginInfo({ ...loginInfo, user: userInfo });
-            });
-          })
-          .catch(() => {
-            alert(ERROR_MESSAGE.MEMBER.FAILURE_PASSWORD_REQUEST);
-          });
-
-        break;
-    }
-  };
 
   const confirmPassword = (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
